+++ src/main/java/sopt/org/umbbaServer/domain/qna/example
@@ -1,0 +1,1 @@
+.
