+++ src/main/java/sopt/org/umbbaServer/domain/user/controller/AuthController.java
@@ -3,15 +3,15 @@ package sopt.org.umbbaServer.domain.user.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import sopt.org.umbbaServer.domain.user.dto.RefreshRequestDto;
-import sopt.org.umbbaServer.domain.user.dto.SocialLoginRequestDto;
-import sopt.org.umbbaServer.domain.user.dto.UserLoginResponseDto;
-import sopt.org.umbbaServer.domain.user.jwt.JwtProvider;
-import sopt.org.umbbaServer.domain.user.jwt.TokenDto;
+import sopt.org.umbbaServer.domain.user.controller.dto.request.RefreshRequestDto;
+import sopt.org.umbbaServer.domain.user.controller.dto.request.SocialLoginRequestDto;
+import sopt.org.umbbaServer.domain.user.controller.dto.response.UserLoginResponseDto;
+import sopt.org.umbbaServer.global.config.jwt.JwtProvider;
+import sopt.org.umbbaServer.global.config.jwt.TokenDto;
 import sopt.org.umbbaServer.domain.user.service.AuthService;
 import sopt.org.umbbaServer.domain.user.social.kakao.KakaoLoginService;
-import sopt.org.umbbaServer.error.ApiResponse;
-import sopt.org.umbbaServer.error.SuccessType;
+import sopt.org.umbbaServer.global.common.dto.ApiResponse;
+import sopt.org.umbbaServer.global.exception.SuccessType;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.Principal;
