+++ src/main/java/sopt/org/umbbaServer/domain/user/controller/dto/request/RefreshRequestDto.java
@@ -1,0 +1,12 @@
+package sopt.org.umbbaServer.domain.user.controller.dto.request;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class RefreshRequestDto {
+
+    private Long userId;
+}
