+++ src/main/java/sopt/org/umbbaServer/domain/user/controller/dto/request/SocialLoginRequestDto.java
@@ -1,0 +1,13 @@
+package sopt.org.umbbaServer.domain.user.controller.dto.request;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import sopt.org.umbbaServer.domain.user.social.SocialPlatform;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class SocialLoginRequestDto {
+
+    private SocialPlatform socialPlatform;
+}
