--- src/main/java/sopt/org/umbbaServer/domain/user/dto/RefreshRequestDto.java
@@ -1,12 +1,0 @@
-package sopt.org.umbbaServer.domain.user.dto;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class RefreshRequestDto {
-
-    private Long userId;
-}
