--- src/main/java/sopt/org/umbbaServer/domain/user/dto/SocialLoginRequestDto.java
@@ -1,13 +1,0 @@
-package sopt.org.umbbaServer.domain.user.dto;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import sopt.org.umbbaServer.domain.user.social.SocialPlatform;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class SocialLoginRequestDto {
-
-    private SocialPlatform socialPlatform;
-}
