--- src/main/java/sopt/org/umbbaServer/domain/user/jwt/JwtValidationType.java
@@ -1,10 +1,0 @@
-package sopt.org.umbbaServer.domain.user.jwt;
-
-public enum JwtValidationType {
-    VALID_JWT,
-    INVALID_JWT_SIGNATURE,
-    INVALID_JWT_TOKEN,
-    EXPIRED_JWT_TOKEN,
-    UNSUPPORTED_JWT_TOKEN,
-    EMPTY_JWT
-}
