--- src/main/java/sopt/org/umbbaServer/domain/user/jwt/TokenDto.java
@@ -1,18 +1,0 @@
-package sopt.org.umbbaServer.domain.user.jwt;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenDto {
-
-    private String accessToken;
-    private String refreshToken;
-
-    public static TokenDto of(String accessToken, String refreshToken) {
-        return new TokenDto(accessToken, refreshToken);
-    }
-}
