--- src/main/java/sopt/org/umbbaServer/domain/user/jwt/redis/TokenRepository.java
@@ -1,7 +1,0 @@
-package sopt.org.umbbaServer.domain.user.jwt.redis;
-
-import org.springframework.data.repository.CrudRepository;
-
-//Redis에 저장해주는 역할
-public interface TokenRepository extends CrudRepository<RefreshToken, Long> {
-}
