--- src/main/java/sopt/org/umbbaServer/domain/user/jwt/security/CustomJwtAuthenticationEntryPoint.java
@@ -1,30 +1,0 @@
-package sopt.org.umbbaServer.domain.user.jwt.security;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-import sopt.org.umbbaServer.error.ApiResponse;
-import sopt.org.umbbaServer.error.ErrorType;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class CustomJwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-      setResponse(response);
-    }
-
-    private void setResponse(HttpServletResponse response) throws IOException {
-        response.setCharacterEncoding("UTF-8");
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        response.getWriter().println(objectMapper.writeValueAsString(ApiResponse.error(ErrorType.INVALID_ACCESS_TOKEN)));
-    }
-}
