--- src/main/java/sopt/org/umbbaServer/domain/user/jwt/security/JwtAuthenticationFilter.java
@@ -1,58 +1,0 @@
-package sopt.org.umbbaServer.domain.user.jwt.security;
-
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-import sopt.org.umbbaServer.domain.user.jwt.JwtProvider;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-import static sopt.org.umbbaServer.domain.user.jwt.JwtValidationType.VALID_JWT;
-
-/**
- * JWT가 유효성을 검증하는 Filter
- */
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-    private final JwtProvider jwtProvider;
-
-    @Override
-    protected void doFilterInternal(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response, @NonNull FilterChain filterChain) throws IOException, ServletException {
-        try {
-
-            // Request의 Header에서 JWT 토큰을 String으로 가져옴
-            final String token = getJwtFromRequest(request);
-
-            if (StringUtils.hasText(token) && jwtProvider.validateAccessToken(token) == VALID_JWT) {
-                Long userId = jwtProvider.getUserFromJwt(token);
-                UserAuthentication authentication = new UserAuthentication(userId, null, null);
-                authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
-        } catch (Exception exception) {
-            log.error("error : ", exception);
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-    private String getJwtFromRequest(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring("Bearer ".length());
-        }
-        return null;
-    }
-}
