--- src/main/java/sopt/org/umbbaServer/domain/user/jwt/security/UserAuthentication.java
@@ -1,15 +1,0 @@
-package sopt.org.umbbaServer.domain.user.jwt.security;
-
-
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-
-import java.util.Collection;
-
-public class UserAuthentication extends UsernamePasswordAuthenticationToken {
-
-    public UserAuthentication(Object principal, Object credentials, Collection<? extends GrantedAuthority> authorities) {
-        super(principal, credentials, authorities);
-
-    }
-}
