+++ src/main/java/sopt/org/umbbaServer/domain/user/service/AuthService.java
@@ -5,19 +5,18 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sopt.org.umbbaServer.domain.user.dto.SocialLoginRequestDto;
-import sopt.org.umbbaServer.domain.user.dto.UserLoginResponseDto;
-import sopt.org.umbbaServer.domain.user.jwt.JwtProvider;
-import sopt.org.umbbaServer.domain.user.jwt.TokenDto;
-import sopt.org.umbbaServer.domain.user.jwt.redis.TokenRepository;
-import sopt.org.umbbaServer.domain.user.jwt.security.UserAuthentication;
+import sopt.org.umbbaServer.domain.user.controller.dto.request.SocialLoginRequestDto;
+import sopt.org.umbbaServer.domain.user.controller.dto.response.UserLoginResponseDto;
+import sopt.org.umbbaServer.global.config.jwt.JwtProvider;
+import sopt.org.umbbaServer.global.config.jwt.TokenDto;
+import sopt.org.umbbaServer.global.config.auth.UserAuthentication;
 import sopt.org.umbbaServer.domain.user.model.User;
 import sopt.org.umbbaServer.domain.user.repository.UserRepository;
 import sopt.org.umbbaServer.domain.user.social.SocialPlatform;
 import sopt.org.umbbaServer.domain.user.social.apple.AppleLoginService;
 import sopt.org.umbbaServer.domain.user.social.kakao.KakaoLoginService;
-import sopt.org.umbbaServer.error.CustomException;
-import sopt.org.umbbaServer.error.ErrorType;
+import sopt.org.umbbaServer.global.exception.CustomException;
+import sopt.org.umbbaServer.global.exception.ErrorType;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
