--- src/main/java/sopt/org/umbbaServer/error/ApiResponse.java
@@ -1,34 +1,0 @@
-package sopt.org.umbbaServer.error;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@Getter
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class ApiResponse<T> {
-
-    private final int code;
-    private final String message;
-    private T data;
-
-    public static ApiResponse success(SuccessType successType) {
-        return new ApiResponse<>(successType.getHttpStatusCode(), successType.getMessage());
-    }
-
-    public static <T> ApiResponse<T> success(SuccessType successType, T data) {
-        return new ApiResponse<T>(successType.getHttpStatusCode(), successType.getMessage(), data);
-    }
-
-    public static ApiResponse error(ErrorType errorType) {
-        return new ApiResponse<>(errorType.getHttpStatusCode(), errorType.getMessage());
-    }
-
-    public static ApiResponse error(ErrorType errorType, String message) {
-        return new ApiResponse<>(errorType.getHttpStatusCode(), message);
-    }
-}
