--- src/main/java/sopt/org/umbbaServer/error/CustomException.java
@@ -1,18 +1,0 @@
-package sopt.org.umbbaServer.error;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-    private final ErrorType errorType;
-
-    public CustomException(ErrorType errorType) {
-        super(errorType.getMessage());
-        this.errorType = errorType;
-    }
-
-    public int getHttpStatus() {
-        return errorType.getHttpStatusCode();
-    }
-}
