+++ src/main/java/sopt/org/umbbaServer/global/common/advice/ControllerExceptionAdvice.java
@@ -1,0 +1,62 @@
+package sopt.org.umbbaServer.global.common.advice;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import sopt.org.umbbaServer.global.common.dto.ApiResponse;
+import sopt.org.umbbaServer.global.exception.CustomException;
+import sopt.org.umbbaServer.global.exception.ErrorType;
+import sopt.org.umbbaServer.global.util.slack.SlackApi;
+
+import javax.validation.UnexpectedTypeException;
+import java.util.Objects;
+
+@RestControllerAdvice
+@Component
+@RequiredArgsConstructor
+public class ControllerExceptionAdvice {
+
+    private final SlackApi slackApi;
+
+    /**
+     * 400 BAD_REQUEST
+     */
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    protected ApiResponse handleMethodArgumentNotValidException(final MethodArgumentNotValidException e) {
+        FieldError fieldError = Objects.requireNonNull(e.getFieldError());
+        return ApiResponse.error(ErrorType.REQUEST_VALIDATION_EXCEPTION, String.format("%s. (%s)", fieldError.getDefaultMessage(), fieldError.getField()));
+    }
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(UnexpectedTypeException.class)
+    protected ApiResponse handleUnexpectedTypeException(final UnexpectedTypeException e) {
+        return ApiResponse.error(ErrorType.VALIDATION_WRONG_TYPE_EXCEPTION);
+    }
+
+    /**
+     * 500 INTERNEL_SERVER
+     */
+//    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+//    @ExceptionHandler(Exception.class)
+//    protected ApiResponse<Object> handleException(final Exception e, final HttpServletRequest request) throws IOException {
+//        //slackApi.sendAlert(e, request);
+//        return ApiResponse.error(ErrorType.INTERNAL_SERVER_ERROR);
+//    }
+
+    /**
+     * CUSTOM_ERROR
+     */
+    @ExceptionHandler(CustomException.class)
+    protected ResponseEntity<ApiResponse> handleSoptException(CustomException e) {
+        return ResponseEntity.status(e.getHttpStatus())
+                .body(ApiResponse.error(e.getErrorType(), e.getMessage()));
+    }
+}
