+++ src/main/java/sopt/org/umbbaServer/global/config/auth/UserAuthentication.java
@@ -1,0 +1,15 @@
+package sopt.org.umbbaServer.global.config.auth;
+
+
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.GrantedAuthority;
+
+import java.util.Collection;
+
+public class UserAuthentication extends UsernamePasswordAuthenticationToken {
+
+    public UserAuthentication(Object principal, Object credentials, Collection<? extends GrantedAuthority> authorities) {
+        super(principal, credentials, authorities);
+
+    }
+}
