+++ src/main/java/sopt/org/umbbaServer/global/config/jwt/JwtValidationType.java
@@ -1,0 +1,10 @@
+package sopt.org.umbbaServer.global.config.jwt;
+
+public enum JwtValidationType {
+    VALID_JWT,
+    INVALID_JWT_SIGNATURE,
+    INVALID_JWT_TOKEN,
+    EXPIRED_JWT_TOKEN,
+    UNSUPPORTED_JWT_TOKEN,
+    EMPTY_JWT
+}
