+++ src/main/java/sopt/org/umbbaServer/global/config/jwt/TokenDto.java
@@ -1,0 +1,18 @@
+package sopt.org.umbbaServer.global.config.jwt;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class TokenDto {
+
+    private String accessToken;
+    private String refreshToken;
+
+    public static TokenDto of(String accessToken, String refreshToken) {
+        return new TokenDto(accessToken, refreshToken);
+    }
+}
