+++ src/main/java/sopt/org/umbbaServer/global/config/jwt/redis/RefreshToken.java
@@ -1,0 +1,29 @@
+package sopt.org.umbbaServer.global.config.jwt.redis;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.redis.core.RedisHash;
+import org.springframework.data.redis.core.TimeToLive;
+
+import javax.persistence.Id;
+import java.util.concurrent.TimeUnit;
+
+@Getter
+@RedisHash("refreshToken")
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RefreshToken {
+
+    @Id
+    @JsonIgnore
+    private Long userId;
+
+    private String refreshToken;
+
+    @TimeToLive(unit = TimeUnit.SECONDS)
+    private Integer expiration;
+}
