+++ src/main/java/sopt/org/umbbaServer/global/config/jwt/redis/TokenRepository.java
@@ -1,0 +1,7 @@
+package sopt.org.umbbaServer.global.config.jwt.redis;
+
+import org.springframework.data.repository.CrudRepository;
+
+//Redis에 저장해주는 역할
+public interface TokenRepository extends CrudRepository<RefreshToken, Long> {
+}
