+++ src/main/java/sopt/org/umbbaServer/global/exception/CustomException.java
@@ -1,0 +1,18 @@
+package sopt.org.umbbaServer.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+    private final ErrorType errorType;
+
+    public CustomException(ErrorType errorType) {
+        super(errorType.getMessage());
+        this.errorType = errorType;
+    }
+
+    public int getHttpStatus() {
+        return errorType.getHttpStatusCode();
+    }
+}
