+++ src/main/java/sopt/org/umbbaServer/global/exception/SuccessType.java
@@ -1,0 +1,31 @@
+package sopt.org.umbbaServer.global.exception;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor(access =  AccessLevel.PRIVATE)
+public enum SuccessType {
+
+    /**
+     * 200 OK
+     */
+    LOGIN_SUCCESS(HttpStatus.OK, "로그인에 성공했습니다."),
+    REISSUE_SUCCESS(HttpStatus.OK, "Access 토큰 재발급에 성공했습니다."),
+    LOGOUT_SUCCESS(HttpStatus.OK, "로그아웃에 성공했습니다."),
+    KAKAO_ACCESS_TOKEN_SUCCESS(HttpStatus.OK, "카카오 엑세스 토큰을 가져오는데 성공했습니다"),
+
+    /**
+     * 201 CREATED
+     */
+    ;
+
+    private final HttpStatus httpStatus;
+    private final String message;
+
+    public int getHttpStatusCode() {
+        return httpStatus.value();
+    }
+}
