--- src/main/java/sopt/org/umbbaServer/util/ServerProfileController.java
@@ -1,22 +1,0 @@
-package sopt.org.umbbaServer.util;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.env.Environment;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Arrays;
-
-@RestController
-@RequiredArgsConstructor
-public class ServerProfileController {
-
-    private final Environment env;
-
-    @GetMapping("/profile")
-    public String getProfile() {
-        return Arrays.stream(env.getActiveProfiles())
-                .findFirst()
-                .orElse("");
-    }
-}
